+++ cucumber-expressions/java/src/main/java/io/cucumber/cucumberexpressions/Expression.java
@@ -4,6 +4,7 @@ import org.apiguardian.api.API;
 
 import java.lang.reflect.Type;
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Pattern;
 
 @API(status = API.Status.STABLE)
@@ -13,4 +14,6 @@ public interface Expression {
     Pattern getRegexp();
 
     String getSource();
+
+    Set<String> getUndefinedParameterTypeNames();
 }