+++ cucumber-expressions/java/src/main/java/io/cucumber/cucumberexpressions/UndefinedParameterTypeExpression.java
@@ -2,13 +2,18 @@ package io.cucumber.cucumberexpressions;
 
 import java.lang.reflect.Type;
 import java.util.List;
+import java.util.Set;
 import java.util.regex.Pattern;
+
+import static java.util.Collections.unmodifiableSet;
 
 final class UndefinedParameterTypeExpression implements Expression {
     private final String source;
+    private final Set<String> undefinedParameterTypeNames;
 
-    UndefinedParameterTypeExpression(String source) {
+    UndefinedParameterTypeExpression(String source, Set<String> undefinedParameterTypeNames) {
         this.source = source;
+        this.undefinedParameterTypeNames = undefinedParameterTypeNames;
     }
 
     @Override
@@ -24,5 +29,10 @@ final class UndefinedParameterTypeExpression implements Expression {
     @Override
     public String getSource() {
         return source;
+    }
+
+    @Override
+    public Set<String> getUndefinedParameterTypeNames() {
+        return unmodifiableSet(undefinedParameterTypeNames);
     }
 }