+++ cucumber-expressions/javascript/src/UndefinedParameterTypeExpression.ts
@@ -2,7 +2,10 @@ import Expression from './Expression'
 import Argument from './Argument'
 
 export default class UndefinedParameterTypeExpression implements Expression {
-  constructor(public readonly source: string) {}
+  constructor(
+    public readonly source: string,
+    public readonly undefinedParameterTypeNames: Set<string>
+  ) {}
 
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
   public match(text: string): Array<Argument<any>> {